Overview and ProfileCreaton class changes
Cleaned up the code in the Overview Panel class and the ProfileCreation class. Got rid of leftover code and documented necessary methods and variables. Made sure to get rid of any warnings that were present as well. This included getting rid of variables that were never used along with adding suppressions where necessary.